+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.killersolyom.projekt">
+    package="ro.sapientia.ms.sapinews">
 
     <application
         android:allowBackup="true"
