+++ app/src/main/java/ro/sapientia/ms/sapinews/Advertisment.java
@@ -1,6 +1,4 @@
-package com.example.killersolyom.projekt;
-
-import android.net.Uri;
+package ro.sapientia.ms.sapinews;
 
 public class Advertisment {
