+++ app/src/main/java/ro/sapientia/ms/sapinews/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.killersolyom.projekt;
+package ro.sapientia.ms.sapinews;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,7 @@ import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 
+import ro.sapientia.ms.sapinews.R;
 public class MainActivity extends AppCompatActivity {
 
     ImageView full_logo;
