+++ app/src/main/java/ro/sapientia/ms/sapinews/User.java
@@ -1,6 +1,4 @@
-package com.example.killersolyom.projekt;
-
-import android.net.Uri;
+package ro.sapientia.ms.sapinews;
 
 public class User {
 
@@ -18,16 +16,7 @@ public class User {
     private User(){
 
     }
-/*
-    public User(String firstName, String lastName, String phoneNumb, String emailAddress, String address){
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.phoneNumb = phoneNumb;
-        this.emailAddress = emailAddress;
-        this.address = address;
-        this.ID=0;
-    }
-*/
+
 
     public static User getInstance(){
         if(single_instance == null){
