+++ app/src/main/res/layout/main_screen_activity.xml
@@ -9,7 +9,7 @@
     android:orientation="vertical"
     tools:context=".MainScreenActivity">
     <fragment
-        android:name="com.example.killersolyom.projekt.HomeFragment"
+        android:name="ro.sapientia.ms.sapinews.HomeFragment"
         android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="0dp"
