+++ app/src/main/res/layout/recyclerview_advertisment.xml
@@ -12,6 +12,7 @@
 
     <ImageView
         android:id="@+id/AdvertismentImage"
+        android:clickable="true"
         android:layout_width="match_parent"
         android:layout_height="150dp"
         android:layout_marginStart="0dp"
