+++ app/src/main/res/values/strings.xml
@@ -21,7 +21,5 @@
     <string name="Address">Address</string>
     <string name="Email">Email</string>
     <string name="Image">Image</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
